--- proyecto/src/main/java/com/clasepdn/proyecto/ProyectoApplication.java
@@ -1,13 +1,0 @@
-package com.clasepdn.proyecto;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProyectoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ProyectoApplication.class, args);
-	}
-
-}
