--- proyecto/src/test/java/com/clasepdn/proyecto/ProyectoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.clasepdn.proyecto;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProyectoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
