+++ src/main/java/com/parcial/pdn_clases_parcial/PdnClasesParcialApplication.java
@@ -1,0 +1,13 @@
+package com.parcial.pdn_clases_parcial;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PdnClasesParcialApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PdnClasesParcialApplication.class, args);
+	}
+
+}
