+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.application.name=pdn_clases_parcial
+#http://localhost:8080/swagger-ui/index.html
