+++ src/test/java/com/parcial/pdn_clases_parcial/PdnClasesParcialApplicationTests.java
@@ -1,0 +1,13 @@
+package com.parcial.pdn_clases_parcial;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PdnClasesParcialApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
